rename 'plot topomap evoked' -> 'plot evoked topomaps'
