+++ mnelab/dialogs/__init__.py
@@ -18,8 +18,7 @@ from .interpolate_bads import InterpolateBadsDialog
 from .meta_info import MetaInfoDialog
 from .montage import MontageDialog
 from .pick_channels import PickChannelsDialog
-from .plot_evoked import PlotEvokedComparisonDialog, PlotEvokedDialog
-from .plot_topomap import PlotTopomapEvokedDialog
+from .plot_evoked import PlotEvokedComparisonDialog, PlotEvokedDialog, PlotEvokedTopomaps
 from .reference import ReferenceDialog
 from .run_ica import RunICADialog
 from .xdf_chunks import XDFChunksDialog
